+++ lib/src/app.dart
@@ -5,6 +5,7 @@ import 'dashboard/dashboard_controller.dart';
 import 'dashboard/dashboard_view.dart';
 import 'jira/jira_controller.dart';
 import 'jira/jira_view.dart';
+import 'jira/multi_task/multi_task_view.dart';
 import 'settings/settings_controller.dart';
 import 'settings/settings_view.dart';
 import 'widgets/update_checker.dart';
@@ -48,11 +49,19 @@ class _MyAppState extends State<MyApp> {
           theme: ThemeData(
             useMaterial3: true,
             fontFamily: 'JetBrainsMono',
+            colorScheme: ColorScheme.fromSeed(
+              seedColor: widget.settingsController.seedColor,
+              brightness: Brightness.light,
+            ),
           ),
-          darkTheme: ThemeData.dark(useMaterial3: true).copyWith(
-              textTheme: ThemeData.dark()
-                  .textTheme
-                  .apply(fontFamily: 'JetBrainsMono')),
+          darkTheme: ThemeData(
+            useMaterial3: true,
+            fontFamily: 'JetBrainsMono',
+            colorScheme: ColorScheme.fromSeed(
+              seedColor: widget.settingsController.seedColor,
+              brightness: Brightness.dark,
+            ),
+          ),
           themeMode: widget.settingsController.themeMode,
           onGenerateRoute: (RouteSettings routeSettings) {
             return MaterialPageRoute<void>(
@@ -66,6 +75,10 @@ class _MyAppState extends State<MyApp> {
                   case DashboardView.routeName:
                     view =
                         DashboardView(controller: widget.dashboardController);
+                    break;
+                  case MultiTaskView.routeName:
+                    view =
+                        MultiTaskView(controller: widget.jiraController);
                     break;
                   default:
                     view = JiraView(controller: widget.jiraController);